# igor-code/channelsconnect/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/listings": {
      "post": {
        "operationId": "ListingsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingEntity"
                }
              }
            }
          }
        },
        "tags": [
          "listings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ListingsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListingEntity"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "listings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/listings/active": {
      "get": {
        "operationId": "ListingsController_findActive",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListingEntity"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "listings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/listings/my-listings": {
      "get": {
        "operationId": "ListingsController_findMyListings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListingEntity"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "listings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/listings/{id}": {
      "get": {
        "operationId": "ListingsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingEntity"
                }
              }
            }
          }
        },
        "tags": [
          "listings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ListingsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingEntity"
                }
              }
            }
          }
        },
        "tags": [
          "listings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ListingsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingEntity"
                }
              }
            }
          }
        },
        "tags": [
          "listings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/bookings/listing/{listingId}": {
      "get": {
        "operationId": "BookingsController_findByListingId",
        "parameters": [
          {
            "name": "listingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all bookings for a specific listing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingEntity"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "bookings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/reports/revenue": {
      "get": {
        "operationId": "ReportsController_getRevenueStats",
        "summary": "Get revenue statistics",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for the report range",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for the report range",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "listingId",
            "required": false,
            "in": "query",
            "description": "Filter by specific listing ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revenue statistics including totals, by month, by listing, and by source"
          }
        },
        "tags": [
          "reports"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reports/occupancy": {
      "get": {
        "operationId": "ReportsController_getOccupancyStats",
        "summary": "Get occupancy statistics",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for the report range",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for the report range",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "listingId",
            "required": false,
            "in": "query",
            "description": "Filter by specific listing ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Occupancy rates by listing and by month"
          }
        },
        "tags": [
          "reports"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reports/adr": {
      "get": {
        "operationId": "ReportsController_getADRStats",
        "summary": "Get Average Daily Rate statistics",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for the report range",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for the report range",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "listingId",
            "required": false,
            "in": "query",
            "description": "Filter by specific listing ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ADR statistics including RevPAR, by month, by listing, and by source"
          }
        },
        "tags": [
          "reports"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reports/payouts": {
      "get": {
        "operationId": "ReportsController_getPayoutStats",
        "summary": "Get payout status statistics",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for the report range",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for the report range",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "listingId",
            "required": false,
            "in": "query",
            "description": "Filter by specific listing ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payout statistics by status and source, with recent bookings"
          }
        },
        "tags": [
          "reports"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Channels Connect API",
    "description": "The Channels Connect API - Property Management & Channel Management Platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateListingDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "propertyType": {
            "type": "string"
          },
          "bedrooms": {
            "type": "number"
          },
          "bathrooms": {
            "type": "number"
          },
          "maxGuests": {
            "type": "number"
          },
          "basePrice": {
            "type": "number"
          },
          "currency": {
            "type": "string",
            "default": "USD"
          },
          "amenities": {
            "type": "object"
          },
          "houseRules": {
            "type": "string"
          },
          "cancellationPolicy": {
            "type": "string"
          },
          "checkInTime": {
            "type": "string"
          },
          "checkOutTime": {
            "type": "string"
          },
          "minNights": {
            "type": "number",
            "default": 1
          },
          "maxNights": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "title"
        ]
      },
      "i": {
        "type": "object",
        "properties": {}
      },
      "ListingEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "userId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/i"
              }
            ]
          },
          "longitude": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/i"
              }
            ]
          },
          "propertyType": {
            "type": "string",
            "nullable": true
          },
          "bedrooms": {
            "type": "number",
            "nullable": true
          },
          "bathrooms": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/i"
              }
            ]
          },
          "beds": {
            "type": "number",
            "nullable": true
          },
          "maxGuests": {
            "type": "number",
            "nullable": true
          },
          "basePrice": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/i"
              }
            ]
          },
          "currency": {
            "type": "string"
          },
          "amenities": {
            "type": "object",
            "nullable": true
          },
          "houseRules": {
            "type": "string",
            "nullable": true
          },
          "cancellationPolicy": {
            "type": "string",
            "nullable": true
          },
          "checkInTime": {
            "type": "string",
            "nullable": true
          },
          "checkOutTime": {
            "type": "string",
            "nullable": true
          },
          "minNights": {
            "type": "number"
          },
          "maxNights": {
            "type": "number",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "beds24PropId": {
            "type": "string",
            "nullable": true,
            "description": "Beds24 property ID"
          },
          "beds24RoomId": {
            "type": "string",
            "nullable": true,
            "description": "Beds24 room/roomType ID"
          },
          "airbnbListingId": {
            "type": "string",
            "nullable": true,
            "description": "Airbnb listing ID"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "title",
          "currency",
          "minNights",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateListingDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "propertyType": {
            "type": "string"
          },
          "bedrooms": {
            "type": "number"
          },
          "bathrooms": {
            "type": "number"
          },
          "maxGuests": {
            "type": "number"
          },
          "basePrice": {
            "type": "number"
          },
          "currency": {
            "type": "string",
            "default": "USD"
          },
          "amenities": {
            "type": "object"
          },
          "houseRules": {
            "type": "string"
          },
          "cancellationPolicy": {
            "type": "string"
          },
          "checkInTime": {
            "type": "string"
          },
          "checkOutTime": {
            "type": "string"
          },
          "minNights": {
            "type": "number",
            "default": 1
          },
          "maxNights": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "BookingEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "userId": {
            "type": "string"
          },
          "listingId": {
            "type": "number"
          },
          "guestName": {
            "type": "string"
          },
          "guestEmail": {
            "type": "string",
            "nullable": true
          },
          "guestPhone": {
            "type": "string",
            "nullable": true
          },
          "checkIn": {
            "format": "date-time",
            "type": "string"
          },
          "checkOut": {
            "format": "date-time",
            "type": "string"
          },
          "numGuests": {
            "type": "number"
          },
          "totalPrice": {
            "$ref": "#/components/schemas/i"
          },
          "status": {
            "type": "string"
          },
          "bookingSource": {
            "type": "string",
            "nullable": true
          },
          "externalId": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "listingId",
          "guestName",
          "checkIn",
          "checkOut",
          "numGuests",
          "totalPrice",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}